extract _actuallySend to file
