+++ packages/insomnia-app/app/network/__tests__/network.test.ts
@@ -16,8 +16,9 @@ import {
 import { filterHeaders } from '../../common/misc';
 import { getRenderedRequestAndContext } from '../../common/render';
 import * as models from '../../models';
+import * as networkUtils from '../curl';
 import { DEFAULT_BOUNDARY } from '../multipart';
-import * as networkUtils from '../network';
+import { _getAwsAuthHeaders, _parseHeaders } from '../network';
 
 const getRenderedRequest = async (args: Parameters<typeof getRenderedRequestAndContext>[0]) => (await getRenderedRequestAndContext(args)).request;
 
@@ -816,7 +817,7 @@ describe('_getAwsAuthHeaders', () => {
       sessionToken: req.authentication.sessionToken || '',
     };
 
-    const headers = networkUtils._getAwsAuthHeaders(
+    const headers = _getAwsAuthHeaders(
       credentials,
       req.headers,
       req.body.text,
@@ -850,7 +851,7 @@ describe('_getAwsAuthHeaders', () => {
       sessionToken: req.authentication.sessionToken || '',
     };
 
-    const headers = networkUtils._getAwsAuthHeaders(
+    const headers = _getAwsAuthHeaders(
       credentials,
       req.headers,
       null,
@@ -886,7 +887,7 @@ describe('_parseHeaders', () => {
   const minimalHeaders = ['HTTP/1.1 301', ''];
 
   it('Parses single response headers', () => {
-    expect(networkUtils._parseHeaders(Buffer.from(basicHeaders.join('\n')))).toEqual([
+    expect(_parseHeaders(Buffer.from(basicHeaders.join('\n')))).toEqual([
       {
         code: 301,
         version: 'HTTP/1.1',
@@ -934,7 +935,7 @@ describe('_parseHeaders', () => {
   });
 
   it('Parses Windows newlines', () => {
-    expect(networkUtils._parseHeaders(Buffer.from(basicHeaders.join('\r\n')))).toEqual([
+    expect(_parseHeaders(Buffer.from(basicHeaders.join('\r\n')))).toEqual([
       {
         code: 301,
         version: 'HTTP/1.1',
@@ -983,7 +984,7 @@ describe('_parseHeaders', () => {
 
   it('Parses multiple responses', () => {
     const blobs = basicHeaders.join('\r\n') + '\n' + minimalHeaders.join('\n');
-    expect(networkUtils._parseHeaders(Buffer.from(blobs))).toEqual([
+    expect(_parseHeaders(Buffer.from(blobs))).toEqual([
       {
         code: 301,
         version: 'HTTP/1.1',
